requirements.txt
Django==4.2.16
djangorestframework==3.15.2
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
drf-yasg==1.21.7
python-decouple==3.8
PyJWT==2.9.0
python-dotenv==1.0.1
gunicorn==23.0.0